+++ tests/django/test_django.py
@@ -38,6 +38,7 @@ def test_middleware_should_call_apilytics_api(
         "requestSize",
         "responseSize",
         "timeMillis",
+        *(("memoryUsage", "memoryTotal") if platform.system() == "Linux" else ()),
     }
     assert data["path"] == "/"
     assert data["method"] == "GET"
@@ -45,6 +46,9 @@ def test_middleware_should_call_apilytics_api(
     assert data["requestSize"] == 0
     assert data["responseSize"] > 0
     assert isinstance(data["timeMillis"], int)
+    if platform.system() == "Linux":
+        assert isinstance(data["memoryUsage"], int)
+        assert isinstance(data["memoryTotal"], int)
 
 
 def test_middleware_should_send_query_params(
@@ -123,12 +127,16 @@ def test_middleware_should_work_with_streaming_response(
         "statusCode",
         "requestSize",
         "timeMillis",
+        *(("memoryUsage", "memoryTotal") if platform.system() == "Linux" else ()),
     }
     assert data["path"] == "/streaming"
     assert data["method"] == "GET"
     assert data["statusCode"] == 200
     assert data["requestSize"] == 0
     assert isinstance(data["timeMillis"], int)
+    if platform.system() == "Linux":
+        assert isinstance(data["memoryUsage"], int)
+        assert isinstance(data["memoryTotal"], int)
 
 
 @django.test.override_settings(APILYTICS_API_KEY=None)
@@ -160,6 +168,7 @@ def test_middleware_should_send_data_even_on_errors(
         "statusCode",
         "requestSize",
         "responseSize",
+        *(("memoryUsage", "memoryTotal") if platform.system() == "Linux" else ()),
     }
     assert data["method"] == "GET"
     assert data["path"] == "/error"
@@ -167,3 +176,6 @@ def test_middleware_should_send_data_even_on_errors(
     assert data["requestSize"] == 0
     assert data["responseSize"] > 0
     assert isinstance(data["timeMillis"], int)
+    if platform.system() == "Linux":
+        assert isinstance(data["memoryUsage"], int)
+        assert isinstance(data["memoryTotal"], int)