+++ tests/test_core.py
@@ -1,4 +1,5 @@
 import platform
+import textwrap
 import unittest.mock
 import urllib.error
 
@@ -34,14 +35,23 @@ def test_apilytics_sender_should_call_apilytics_api(
     }
 
     data = tests.conftest.decode_request_data(call_kwargs["data"])
-    assert data.keys() == {"path", "method", "statusCode", "timeMillis"}
+    assert data.keys() == {
+        "path",
+        "method",
+        "statusCode",
+        "timeMillis",
+        *(("memoryUsage", "memoryTotal") if platform.system() == "Linux" else ()),
+    }
     assert data["path"] == "/"
     assert data["method"] == "GET"
     assert data["statusCode"] == 200
     assert isinstance(data["timeMillis"], int)
+    if platform.system() == "Linux":
+        assert data["memoryUsage"] > 0
+        assert data["memoryTotal"] > data["memoryUsage"]
 
 
-def test_middleware_should_send_query_params(
+def test_apilytics_sender_should_send_query_params(
     mocked_urlopen: unittest.mock.MagicMock,
 ) -> None:
     with apilytics.core.ApilyticsSender(
@@ -59,7 +69,7 @@ def test_middleware_should_send_query_params(
     assert data["query"] == "key=value?other=123"
 
 
-def test_middleware_should_not_send_empty_query_params(
+def test_apilytics_sender_should_not_send_empty_query_params(
     mocked_urlopen: unittest.mock.MagicMock,
 ) -> None:
     with apilytics.core.ApilyticsSender(
@@ -107,10 +117,142 @@ def test_apilytics_sender_should_handle_empty_values_correctly(
     assert mocked_urlopen.call_count == 1
     __, call_kwargs = mocked_urlopen.call_args
     data = tests.conftest.decode_request_data(call_kwargs["data"])
-    assert data.keys() == {"path", "method", "timeMillis"}
+    assert data.keys() == {
+        "path",
+        "method",
+        "timeMillis",
+        *(("memoryUsage", "memoryTotal") if platform.system() == "Linux" else ()),
+    }
     assert data["path"] == ""
     assert data["method"] == ""
     assert isinstance(data["timeMillis"], int)
+    if platform.system() == "Linux":
+        assert isinstance(data["memoryUsage"], int)
+        assert isinstance(data["memoryTotal"], int)
+
+
+@unittest.mock.patch("apilytics.core.platform.system", return_value="Linux")
+def test_apilytics_sender_should_read_proc_meminfo_on_linux(
+    _mocked_system: unittest.mock.MagicMock,
+    mocked_urlopen: unittest.mock.MagicMock,
+) -> None:
+    memory_total = 4_125_478_912
+    memory_available = 3_360_526_336
+
+    mocked_meminfo = textwrap.dedent(
+        f"""\
+        MemTotal:        {memory_total // 1024} kB
+        MemFree:          789940 kB
+        MemAvailable:    {memory_available // 1024} kB
+        Buffers:         2450168 kB
+        """  # The real file is longer.
+    )
+    with unittest.mock.patch(
+        "builtins.open", new=unittest.mock.mock_open(read_data=mocked_meminfo)
+    ) as mocked_open:
+        with apilytics.core.ApilyticsSender(
+            api_key="dummy-key",
+            path="/",
+            method="GET",
+        ) as sender:
+            sender.set_response_info(status_code=200)
+
+    assert mocked_open.call_count == 1
+    assert mocked_urlopen.call_count == 1
+    __, call_kwargs = mocked_urlopen.call_args
+    data = tests.conftest.decode_request_data(call_kwargs["data"])
+    assert data["memoryUsage"] == memory_total - memory_available
+    assert data["memoryTotal"] == memory_total
+
+
+@unittest.mock.patch("apilytics.core.platform.system", return_value="Linux")
+def test_apilytics_sender_should_handle_proc_meminfo_read_failure(
+    _mocked_system: unittest.mock.MagicMock,
+    mocked_urlopen: unittest.mock.MagicMock,
+) -> None:
+    with unittest.mock.patch("builtins.open", side_effect=OSError) as mocked_open:
+        with apilytics.core.ApilyticsSender(
+            api_key="dummy-key",
+            path="/",
+            method="GET",
+        ) as sender:
+            sender.set_response_info(status_code=200)
+
+    assert mocked_open.call_count == 1
+    assert mocked_urlopen.call_count == 1
+    __, call_kwargs = mocked_urlopen.call_args
+    data = tests.conftest.decode_request_data(call_kwargs["data"])
+    assert "memoryUsage" not in data
+    assert "memoryTotal" not in data
+
+
+@unittest.mock.patch("apilytics.core.platform.system", return_value="Linux")
+def test_apilytics_sender_should_handle_proc_meminfo_total_missing(
+    _mocked_system: unittest.mock.MagicMock,
+    mocked_urlopen: unittest.mock.MagicMock,
+) -> None:
+    with unittest.mock.patch(
+        "builtins.open", new=unittest.mock.mock_open(read_data="")
+    ) as mocked_open:
+        with apilytics.core.ApilyticsSender(
+            api_key="dummy-key",
+            path="/",
+            method="GET",
+        ) as sender:
+            sender.set_response_info(status_code=200)
+
+    assert mocked_open.call_count == 1
+    assert mocked_urlopen.call_count == 1
+    __, call_kwargs = mocked_urlopen.call_args
+    data = tests.conftest.decode_request_data(call_kwargs["data"])
+    assert "memoryUsage" not in data
+    assert "memoryTotal" not in data
+
+
+@unittest.mock.patch("apilytics.core.platform.system", return_value="Linux")
+def test_apilytics_sender_should_handle_proc_meminfo_available_missing(
+    _mocked_system: unittest.mock.MagicMock,
+    mocked_urlopen: unittest.mock.MagicMock,
+) -> None:
+    memory_total = 1048576
+    with unittest.mock.patch(
+        "builtins.open",
+        new=unittest.mock.mock_open(read_data=f"MemTotal: {memory_total // 1024}"),
+    ) as mocked_open:
+        with apilytics.core.ApilyticsSender(
+            api_key="dummy-key",
+            path="/",
+            method="GET",
+        ) as sender:
+            sender.set_response_info(status_code=200)
+
+    assert mocked_open.call_count == 1
+    assert mocked_urlopen.call_count == 1
+    __, call_kwargs = mocked_urlopen.call_args
+    data = tests.conftest.decode_request_data(call_kwargs["data"])
+    assert "memoryUsage" not in data
+    assert data["memoryTotal"] == memory_total
+
+
+@unittest.mock.patch("apilytics.core.platform.system", return_value="Windows")
+def test_apilytics_sender_should_not_read_proc_meminfo_when_not_on_linux(
+    _mocked_system: unittest.mock.MagicMock,
+    mocked_urlopen: unittest.mock.MagicMock,
+) -> None:
+    with unittest.mock.patch("builtins.open") as mocked_open:
+        with apilytics.core.ApilyticsSender(
+            api_key="dummy-key",
+            path="/",
+            method="GET",
+        ) as sender:
+            sender.set_response_info(status_code=200)
+
+    assert mocked_open.call_count == 0
+    assert mocked_urlopen.call_count == 1
+    __, call_kwargs = mocked_urlopen.call_args
+    data = tests.conftest.decode_request_data(call_kwargs["data"])
+    assert "memoryUsage" not in data
+    assert "memoryTotal" not in data
 
 
 @unittest.mock.patch(