+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Send current system memory usage and total available memory together with metrics on Linux systems.
+
 ## [1.3.0] - 2022-02-02
 
 ### Added